Move spec packages in their parent package
