+++ web/apps/spec/apps_test.go
@@ -1,7 +1,7 @@
 // apps_test package is introduced to avoid circular dependencies since this
 // particular test requires to depend on routing directly to expose the API
 // and the APP server.
-package appsspec
+package spec
 
 import (
 	"bytes"
@@ -225,7 +225,7 @@ func TestServeAppsWithACode(t *testing.T) {
 
 func TestMain(m *testing.M) {
 	config.UseTestFile()
-	config.GetConfig().Assets = "../../assets"
+	config.GetConfig().Assets = "../../../assets"
 
 	tempdir, err := ioutil.TempDir("", "cozy-stack")
 	if err != nil {
