--- web/appsspec/empty.go
@@ -1,1 +1,0 @@
-package appsspec
