+++ web/auth/spec/empty.go
@@ -1,0 +1,1 @@
+package spec
