--- web/authspec/empty.go
@@ -1,1 +1,0 @@
-package authspec
